a2a_langgraph_demo/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5
langchain-core>=0.1.0
langgraph>=0.0.20
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.4.2
openai>=1.3.0
google-generativeai>=0.3.0
requests>=2.31.0
sse-starlette>=1.6.5
python-multipart>=0.0.6
sentry-sdk>=1.40.0
sseclient-py>=1.7.2
